feat: CastData with PlaylistId - 북마크 기능을 위해서 CastData에 PlaylistId를 넣었습니다

+++ app/src/main/java/kr/dori/android/own_cast/forApiData/AuthRetrofitInterface.kt
@@ -105,7 +105,7 @@ interface Bookmark{
             Response<AuthResponse<Long>>
 
     @GET("/api/study/{playlistId}")
-    suspend fun getBookmark(@Path("playlistId")playlistId: Int):
+    suspend fun getBookmark(@Path("playlistId")playlistId: Long):
             Response<AuthResponse<List<GetBookmark>>>
 
     @GET("/api/study/savedcast")
